updoot

+++ src/commands/sanitize.js
@@ -13,6 +13,7 @@ exports.func = async (res, msg, args) => {
         await messages[0].delete();
         return res.success('Purged last message.', msg.author);
     }
+    if(!msg.channel.permissionsFor(msg.client.user).hasPermission('MANAGE_MESSAGES')) return res.error('I need `Manage Messages` permissions to sanitize.');
     const deleted = await msg.channel.bulkDelete(messages);
     return res.success(`Purged last ${deleted.size} messages.`, msg.author);
 };
